+++ medusa/indexers/indexer_config.py
@@ -4,7 +4,7 @@
 
 from medusa.app import BASE_PYMEDUSA_URL
 from medusa.indexers.tmdb.tmdb import Tmdb
-from medusa.indexers.tvdbv2.tvdbv2_api import TVDBv2
+from medusa.indexers.tvdb.api import TVDBv2
 from medusa.indexers.tvmaze.tvmaze_api import TVmaze
 from medusa.session.core import MedusaSession
